backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
pydantic-settings==2.2.1

# Data & DSL stack
numpy==1.26.4
pandas==2.2.2
pyarrow==15.0.2
duckdb==1.0.0
pyyaml==6.0.1
typer==0.12.3

# Dev / tests
pytest==8.2.1
rich==13.7.1
